musicapi/apps/music/views.py
from rest_framework import generics
from .models import Songs
from .serializers import SongsSerializer


class CreateSongsView(generics.CreateAPIView):
    """
    Provides a get method handler.
    """
    queryset = Songs.objects.all()
    serializer_class = SongsSerializer

class DeleteSongsView(generics.DestroyAPIView):
    """
    Provides a get method handler.
    """
    queryset = Songs.objects.all()
    serializer_class = SongsSerializer

class ListSongsView(generics.ListAPIView):
    """
    Provides a get method handler.
    """
    queryset = Songs.objects.all()
    serializer_class = SongsSerializer

