requirements.txt
Django==2.0.3
djangorestframework==3.9.1
gunicorn==19.9.0
psycopg2-binary==2.7.7
pytz==2018.9
whitenoise==4.1.2
